requirements/dev_requirements.txt
django-debug-toolbar
ipython
django_extensions
-r base_requirements.txt